+++ tests/test_connection.py
@@ -1,6 +1,8 @@
 """Tests for `ember_mug.mug connections`."""
 from __future__ import annotations
 
+from datetime import datetime
+from typing import TYPE_CHECKING
 from unittest.mock import AsyncMock, Mock, patch
 
 import pytest
@@ -19,6 +21,11 @@ from ember_mug.consts import (
 from ember_mug.data import Colour, Model
 from ember_mug.mug import EmberMug
 
+if TYPE_CHECKING:
+
+    class MockMug(EmberMug):
+        _client: AsyncMock
+
 
 @patch("ember_mug.mug.IS_LINUX", True)
 async def test_adapter_with_bluez(ble_device: BLEDevice):
@@ -35,9 +42,9 @@ async def test_adapter_without_bluez(ble_device: BLEDevice):
 @patch("ember_mug.mug.EmberMug.subscribe")
 @patch("ember_mug.mug.establish_connection")
 async def test_connect(
-    mug_subscribe: AsyncMock,
-    mock_establish_connection: AsyncMock,
-    ember_mug: AsyncMock,
+    mug_subscribe: Mock,
+    mock_establish_connection: Mock,
+    ember_mug: MockMug,
 ) -> None:
     # Already connected
     ember_mug._client = AsyncMock()
@@ -48,29 +55,29 @@ async def test_connect(
     mock_establish_connection.assert_not_called()
 
     # Not connected
-    ember_mug._client = None
-    ember_mug.disconnect = AsyncMock()
-    async with ember_mug.connection():
-        pass
+    mock_disconnect = AsyncMock()
+    with patch.multiple(ember_mug, _client=None, disconnect=mock_disconnect):
+        async with ember_mug.connection():
+            pass
 
-    mock_establish_connection.assert_called()
-    mug_subscribe.assert_called()
-    assert ember_mug._client is not None
-    ember_mug.disconnect.assert_called()
+        mock_establish_connection.assert_called()
+        mug_subscribe.assert_called()
+        assert ember_mug._client is not None
+        mock_disconnect.assert_called()
 
 
 @patch("ember_mug.mug.logger")
 @patch("ember_mug.mug.establish_connection")
 async def test_connect_error(
-    mock_establish_connection: AsyncMock,
+    mock_establish_connection: Mock,
     mock_logger: Mock,
-    ember_mug: AsyncMock,
+    ember_mug: MockMug,
 ) -> None:
     ember_mug._client = None  # type: ignore[assignment]
     mock_establish_connection.side_effect = BleakError
     with pytest.raises(BleakError):
         await ember_mug._ensure_connection()
-    msg, device, exception = mock_logger.error.mock_calls[0].args
+    msg, device, exception = mock_logger.debug.mock_calls[1].args
     assert msg == "%s: Failed to connect to the mug: %s"
     assert device == ember_mug.device
     assert isinstance(exception, BleakError)
@@ -79,9 +86,9 @@ async def test_connect_error(
 @patch("ember_mug.mug.logger")
 @patch("ember_mug.mug.establish_connection")
 async def test_pairing_exceptions_esphome(
-    mock_establish_connection: AsyncMock,
+    mock_establish_connection: Mock,
     mock_logger: Mock,
-    ember_mug: AsyncMock,
+    ember_mug: MockMug,
 ) -> None:
     ember_mug._client.is_connected = False
     mock_client = AsyncMock()
@@ -104,8 +111,8 @@ async def test_pairing_exceptions_esphome(
 
 @patch("ember_mug.mug.establish_connection")
 async def test_pairing_exceptions(
-    mock_establish_connection: AsyncMock,
-    ember_mug: AsyncMock,
+    mock_establish_connection: Mock,
+    ember_mug: MockMug,
 ) -> None:
     mock_client = AsyncMock()
     mock_client.pair.side_effect = BleakError
@@ -118,7 +125,7 @@ async def test_pairing_exceptions(
         await ember_mug._ensure_connection()
 
 
-async def test_disconnect(ember_mug: AsyncMock) -> None:
+async def test_disconnect(ember_mug: MockMug) -> None:
     mock_client = AsyncMock()
     ember_mug._client = mock_client
 
@@ -135,7 +142,7 @@ async def test_disconnect(ember_mug: AsyncMock) -> None:
 
 
 @patch("ember_mug.mug.logger")
-def test_disconnect_callback(mock_logger: Mock, ember_mug: AsyncMock) -> None:
+def test_disconnect_callback(mock_logger: Mock, ember_mug: MockMug) -> None:
     ember_mug._expected_disconnect = True
     ember_mug._disconnect_callback(AsyncMock())
     mock_logger.debug.assert_called_with("Disconnect callback called")
@@ -149,7 +156,7 @@ def test_disconnect_callback(mock_logger: Mock, ember_mug: AsyncMock) -> None:
 @patch("ember_mug.mug.logger")
 async def test_read(
     mock_logger: Mock,
-    ember_mug: AsyncMock,
+    ember_mug: MockMug,
 ) -> None:
     with patch.object(ember_mug, "_ensure_connection", AsyncMock()):
         ember_mug._client.read_gatt_char.return_value = b"TEST"
@@ -165,7 +172,7 @@ async def test_read(
 
 
 @patch("ember_mug.mug.logger")
-async def test_write(mock_logger: Mock, ember_mug: AsyncMock) -> None:
+async def test_write(mock_logger: Mock, ember_mug: MockMug) -> None:
     with patch.object(ember_mug, "_ensure_connection", AsyncMock()):
         test_name = bytearray(b"TEST")
         await ember_mug._write(
@@ -200,7 +207,7 @@ async def test_write(mock_logger: Mock, ember_mug: AsyncMock) -> None:
         assert isinstance(exception, BleakError)
 
 
-def test_set_device(ember_mug: AsyncMock) -> None:
+def test_set_device(ember_mug: MockMug) -> None:
     new_device = BLEDevice(
         address="BA:36:a5:be:88:cb",
         name="Ember Ceramic Mug",
@@ -212,7 +219,7 @@ def test_set_device(ember_mug: AsyncMock) -> None:
     assert ember_mug.device.address == new_device.address
 
 
-async def test_get_mug_meta(ember_mug: AsyncMock):
+async def test_get_mug_meta(ember_mug: MockMug) -> None:
     with patch.object(ember_mug, "_ensure_connection", AsyncMock()):
         ember_mug._client.read_gatt_char.return_value = b"Yw====-ABCDEFGHIJ"
         meta = await ember_mug.get_meta()
@@ -221,7 +228,7 @@ async def test_get_mug_meta(ember_mug: AsyncMock):
         ember_mug._client.read_gatt_char.assert_called_once_with(MugCharacteristic.MUG_ID.uuid)
 
 
-async def test_get_mug_battery(ember_mug: AsyncMock):
+async def test_get_mug_battery(ember_mug: MockMug) -> None:
     with patch.object(ember_mug, "_ensure_connection", AsyncMock()):
         ember_mug._client.read_gatt_char.return_value = b"5\x01"
         battery = await ember_mug.get_battery()
@@ -230,7 +237,7 @@ async def test_get_mug_battery(ember_mug: AsyncMock):
         ember_mug._client.read_gatt_char.assert_called_once_with(MugCharacteristic.BATTERY.uuid)
 
 
-async def test_get_mug_led_colour(ember_mug: AsyncMock):
+async def test_get_mug_led_colour(ember_mug: MockMug) -> None:
     with patch.object(ember_mug, "_ensure_connection", AsyncMock()):
         ember_mug._client.read_gatt_char.return_value = b"\xf4\x00\xa1\xff"
         colour = await ember_mug.get_led_colour()
@@ -238,120 +245,126 @@ async def test_get_mug_led_colour(ember_mug: AsyncMock):
         ember_mug._client.read_gatt_char.assert_called_once_with(MugCharacteristic.LED.uuid)
 
 
-async def test_set_mug_led_colour(ember_mug: AsyncMock):
-    with patch.object(ember_mug, "_ensure_connection", AsyncMock()):
+async def test_set_mug_led_colour(ember_mug: MockMug) -> None:
+    mock_ensure_connection = AsyncMock()
+    with patch.object(ember_mug, "_ensure_connection", mock_ensure_connection):
         await ember_mug.set_led_colour(Colour(244, 0, 161))
-        ember_mug._ensure_connection.assert_called_once()
+        mock_ensure_connection.assert_called_once()
         ember_mug._client.write_gatt_char.assert_called_once_with(
             MugCharacteristic.LED.uuid,
             bytearray(b"\xf4\x00\xa1\xff"),
         )
 
 
-async def test_set_volume_level_travel_mug(ember_mug: AsyncMock):
+async def test_set_volume_level_travel_mug(ember_mug: MockMug) -> None:
     ember_mug.is_travel_mug = True
-    with patch.object(ember_mug, "_ensure_connection", AsyncMock()):
+    mock_ensure_connection = AsyncMock()
+    with patch.object(ember_mug, "_ensure_connection", mock_ensure_connection):
         await ember_mug.set_volume_level(VolumeLevel.HIGH)
-        ember_mug._ensure_connection.assert_called_once()
+        mock_ensure_connection.assert_called_once()
         ember_mug._client.write_gatt_char.assert_called_once_with(
             MugCharacteristic.VOLUME.uuid,
             bytearray(b"\02"),
         )
-        ember_mug._ensure_connection.reset_mock()
+        mock_ensure_connection.reset_mock()
         ember_mug._client.write_gatt_char.reset_mock()
 
         await ember_mug.set_volume_level(0)
-        ember_mug._ensure_connection.assert_called_once()
+        mock_ensure_connection.assert_called_once()
         ember_mug._client.write_gatt_char.assert_called_once_with(
             MugCharacteristic.VOLUME.uuid,
             bytearray(b"\00"),
         )
 
 
-async def test_set_volume_level_mug(ember_mug: AsyncMock):
-    with patch.object(ember_mug, "_ensure_connection", AsyncMock()):
+async def test_set_volume_level_mug(ember_mug: MockMug) -> None:
+    mock_ensure_connection = AsyncMock()
+    with patch.object(ember_mug, "_ensure_connection", mock_ensure_connection):
         with pytest.raises(NotImplementedError):
             await ember_mug.set_volume_level(VolumeLevel.HIGH)
-        ember_mug._ensure_connection.assert_not_called()
+        mock_ensure_connection.assert_not_called()
         ember_mug._client.write_gatt_char.assert_not_called()
 
 
-async def test_get_mug_target_temp(ember_mug: AsyncMock):
+async def test_get_mug_target_temp(ember_mug: MockMug) -> None:
     with patch.object(ember_mug, "_ensure_connection", AsyncMock()):
         ember_mug._client.read_gatt_char.return_value = b"\xcd\x15"
         assert (await ember_mug.get_target_temp()) == 55.81
         ember_mug._client.read_gatt_char.assert_called_once_with(MugCharacteristic.TARGET_TEMPERATURE.uuid)
 
 
-async def test_set_mug_target_temp(ember_mug: AsyncMock):
-    with patch.object(ember_mug, "_ensure_connection", AsyncMock()):
+async def test_set_mug_target_temp(ember_mug: MockMug) -> None:
+    mock_ensure_connection = AsyncMock()
+    with patch.object(ember_mug, "_ensure_connection", mock_ensure_connection):
         await ember_mug.set_target_temp(55.81)
-        ember_mug._ensure_connection.assert_called_once()
+        mock_ensure_connection.assert_called_once()
         ember_mug._client.write_gatt_char.assert_called_once_with(
             MugCharacteristic.TARGET_TEMPERATURE.uuid,
             bytearray(b"\xcd\x15"),
         )
 
 
-async def test_get_mug_current_temp(ember_mug: AsyncMock):
+async def test_get_mug_current_temp(ember_mug: MockMug) -> None:
     with patch.object(ember_mug, "_ensure_connection", AsyncMock()):
         ember_mug._client.read_gatt_char.return_value = b"\xcd\x15"
         assert (await ember_mug.get_current_temp()) == 55.81
         ember_mug._client.read_gatt_char.assert_called_once_with(MugCharacteristic.CURRENT_TEMPERATURE.uuid)
 
 
-async def test_get_mug_liquid_level(ember_mug: AsyncMock):
+async def test_get_mug_liquid_level(ember_mug: MockMug) -> None:
     with patch.object(ember_mug, "_ensure_connection", AsyncMock()):
         ember_mug._client.read_gatt_char.return_value = b"\n"
         assert (await ember_mug.get_liquid_level()) == 10
         ember_mug._client.read_gatt_char.assert_called_once_with(MugCharacteristic.LIQUID_LEVEL.uuid)
 
 
-async def test_get_mug_liquid_state(ember_mug: AsyncMock):
+async def test_get_mug_liquid_state(ember_mug: MockMug) -> None:
     with patch.object(ember_mug, "_ensure_connection", AsyncMock()):
         ember_mug._client.read_gatt_char.return_value = b"\x06"
         assert (await ember_mug.get_liquid_state()) == 6
         ember_mug._client.read_gatt_char.assert_called_once_with(MugCharacteristic.LIQUID_STATE.uuid)
 
 
-async def test_get_mug_name(ember_mug):
+async def test_get_mug_name(ember_mug: MockMug) -> None:
     with patch.object(ember_mug, "_ensure_connection", AsyncMock()):
         ember_mug._client.read_gatt_char.return_value = b"Mug Name"
         assert (await ember_mug.get_name()) == "Mug Name"
         ember_mug._client.read_gatt_char.assert_called_once_with(MugCharacteristic.MUG_NAME.uuid)
 
 
-async def test_set_mug_name(ember_mug):
+async def test_set_mug_name(ember_mug: MockMug) -> None:
     with patch.object(ember_mug, "_ensure_connection", AsyncMock()), pytest.raises(ValueError):
         await ember_mug.set_name("Hé!")
 
-    with patch.object(ember_mug, "_ensure_connection", AsyncMock()):
+    mock_ensure_connection = AsyncMock()
+    with patch.object(ember_mug, "_ensure_connection", mock_ensure_connection):
         await ember_mug.set_name("Mug name")
-        ember_mug._ensure_connection.assert_called()
+        mock_ensure_connection.assert_called()
         ember_mug._client.write_gatt_char.assert_called_once_with(
             MugCharacteristic.MUG_NAME.uuid,
             bytearray(b"Mug name"),
         )
 
 
-async def test_get_mug_udsk(ember_mug: AsyncMock):
+async def test_get_mug_udsk(ember_mug: MockMug) -> None:
     with patch.object(ember_mug, "_ensure_connection", AsyncMock()):
         ember_mug._client.read_gatt_char.return_value = b"abcd12345"
         assert (await ember_mug.get_udsk()) == "YWJjZDEyMzQ1"
         ember_mug._client.read_gatt_char.assert_called_once_with(MugCharacteristic.UDSK.uuid)
 
 
-async def test_set_mug_udsk(ember_mug: AsyncMock):
-    with patch.object(ember_mug, "_ensure_connection", AsyncMock()):
+async def test_set_mug_udsk(ember_mug: MockMug) -> None:
+    mock_ensure_connection = AsyncMock()
+    with patch.object(ember_mug, "_ensure_connection", mock_ensure_connection):
         await ember_mug.set_udsk("abcd12345")
-        ember_mug._ensure_connection.assert_called_once()
+        mock_ensure_connection.assert_called_once()
         ember_mug._client.write_gatt_char.assert_called_once_with(
             MugCharacteristic.UDSK.uuid,
             bytearray(b"YWJjZDEyMzQ1"),
         )
 
 
-async def test_get_mug_dsk(ember_mug):
+async def test_get_mug_dsk(ember_mug: MockMug) -> None:
     with patch.object(ember_mug, "_ensure_connection", AsyncMock()):
         ember_mug._client.read_gatt_char.return_value = b"abcd12345"
         assert (await ember_mug.get_dsk()) == "YWJjZDEyMzQ1"
@@ -360,7 +373,7 @@ async def test_get_mug_dsk(ember_mug):
         ember_mug._client.read_gatt_char.assert_called_with(MugCharacteristic.DSK.uuid)
 
 
-async def test_get_mug_temperature_unit(ember_mug: AsyncMock):
+async def test_get_mug_temperature_unit(ember_mug: MockMug) -> None:
     with patch.object(ember_mug, "_ensure_connection", AsyncMock()):
         ember_mug._client.read_gatt_char.return_value = b"\x01"
         assert (await ember_mug.get_temperature_unit()) == TemperatureUnit.FAHRENHEIT
@@ -371,44 +384,47 @@ async def test_get_mug_temperature_unit(ember_mug: AsyncMock):
         ember_mug._client.read_gatt_char.assert_called_once_with(MugCharacteristic.TEMPERATURE_UNIT.uuid)
 
 
-async def test_set_mug_temperature_unit(ember_mug: AsyncMock):
-    with patch.object(ember_mug, "_ensure_connection", AsyncMock()):
+async def test_set_mug_temperature_unit(ember_mug: MockMug) -> None:
+    mock_ensure_connection = AsyncMock()
+    with patch.object(ember_mug, "_ensure_connection", mock_ensure_connection):
         await ember_mug.set_temperature_unit(TemperatureUnit.CELSIUS)
-        ember_mug._ensure_connection.assert_called_once()
+        mock_ensure_connection.assert_called_once()
         ember_mug._client.write_gatt_char.assert_called_once_with(
             MugCharacteristic.TEMPERATURE_UNIT.uuid,
             bytearray(b"\x00"),
         )
 
 
-async def test_mug_ensure_correct_unit(ember_mug: AsyncMock):
+async def test_mug_ensure_correct_unit(ember_mug: MockMug) -> None:
     with patch.object(ember_mug, "_ensure_connection", AsyncMock()):
         ember_mug.data.temperature_unit = TemperatureUnit.CELSIUS
         ember_mug.data.use_metric = True
-        ember_mug.set_temperature_unit = AsyncMock(return_value=None)
-        await ember_mug.ensure_correct_unit()
-        ember_mug.set_temperature_unit.assert_not_called()
-        ember_mug.data.temperature_unit = TemperatureUnit.FAHRENHEIT
-        await ember_mug.ensure_correct_unit()
-        ember_mug.set_temperature_unit.assert_called_with(TemperatureUnit.CELSIUS)
+        mock_set_temp = AsyncMock(return_value=None)
+        with patch.object(ember_mug, 'set_temperature_unit', mock_set_temp):
+            await ember_mug.ensure_correct_unit()
+            mock_set_temp.assert_not_called()
+            ember_mug.data.temperature_unit = TemperatureUnit.FAHRENHEIT
+            await ember_mug.ensure_correct_unit()
+            mock_set_temp.assert_called_with(TemperatureUnit.CELSIUS)
 
 
-async def test_get_mug_battery_voltage(ember_mug: AsyncMock):
+async def test_get_mug_battery_voltage(ember_mug: MockMug) -> None:
     with patch.object(ember_mug, "_ensure_connection", AsyncMock()):
         ember_mug._client.read_gatt_char.return_value = b"\x01"
         assert (await ember_mug.get_battery_voltage()) == 1
         ember_mug._client.read_gatt_char.assert_called_once_with(MugCharacteristic.CONTROL_REGISTER_DATA.uuid)
 
 
-async def test_get_mug_date_time_zone(ember_mug: AsyncMock):
+async def test_get_mug_date_time_zone(ember_mug: MockMug) -> None:
     with patch.object(ember_mug, "_ensure_connection", AsyncMock()):
         ember_mug._client.read_gatt_char.return_value = b"c\x0f\xf6\x00"
         date_time = await ember_mug.get_date_time_zone()
+        assert isinstance(date_time, datetime)
         assert date_time.timestamp() == 1661990400.0
         ember_mug._client.read_gatt_char.assert_called_once_with(MugCharacteristic.DATE_TIME_AND_ZONE.uuid)
 
 
-async def test_read_firmware(ember_mug: AsyncMock):
+async def test_read_firmware(ember_mug: MockMug) -> None:
     with patch.object(ember_mug, "_ensure_connection", AsyncMock()):
         ember_mug._client.read_gatt_char.return_value = b"c\x01\x80\x00\x12\x00"
         firmware = await ember_mug.get_firmware()
@@ -418,57 +434,59 @@ async def test_read_firmware(ember_mug: AsyncMock):
         ember_mug._client.read_gatt_char.assert_called_once_with(MugCharacteristic.FIRMWARE.uuid)
 
 
-async def test_mug_update_initial(ember_mug):
+async def test_mug_update_initial(ember_mug: MockMug) -> None:
     no_extra = INITIAL_ATTRS - EXTRA_ATTRS
 
-    with patch.object(ember_mug, "_ensure_connection", AsyncMock()):
-        ember_mug._update_multiple = AsyncMock(return_value={})
+    mock_update = AsyncMock(return_value={})
+    with patch.multiple(ember_mug, _ensure_connection=AsyncMock(), _update_multiple=mock_update):
         ember_mug.data.model = Model(EMBER_MUG, include_extra=False)
         assert (await ember_mug.update_initial()) == {}
-        ember_mug._update_multiple.assert_called_once_with(no_extra)
+        mock_update.assert_called_once_with(no_extra)
 
-    # Try with extra
-    with patch.object(ember_mug, "_ensure_connection", AsyncMock()):
-        ember_mug._update_multiple.reset_mock()
+        # Try with extra
+        mock_update.reset_mock()
         ember_mug.data.model = Model(EMBER_MUG, include_extra=True)
         assert (await ember_mug.update_initial()) == {}
-        ember_mug._update_multiple.assert_called_once_with(INITIAL_ATTRS)
+        mock_update.assert_called_once_with(INITIAL_ATTRS)
 
 
-async def test_mug_update_all(ember_mug):
-    with patch.object(ember_mug, "_ensure_connection", AsyncMock()):
-        ember_mug._update_multiple = AsyncMock(return_value={})
+async def test_mug_update_all(ember_mug: MockMug) -> None:
+    mock_update = AsyncMock(return_value={})
+    with patch.multiple(ember_mug, _ensure_connection=AsyncMock(), _update_multiple=mock_update):
         assert (await ember_mug.update_all()) == {}
-        ember_mug._update_multiple.assert_called_once_with(UPDATE_ATTRS - EXTRA_ATTRS)
+        mock_update.assert_called_once_with(UPDATE_ATTRS - EXTRA_ATTRS)
 
-    # Try with extras
-    with patch.object(ember_mug, "_ensure_connection", AsyncMock()):
-        ember_mug._update_multiple.reset_mock()
-        ember_mug._update_attrs = UPDATE_ATTRS
+        # Try with extras
+        mock_update.reset_mock()
         assert (await ember_mug.update_all()) == {}
-        ember_mug._update_multiple.assert_called_once_with(UPDATE_ATTRS)
+        mock_update.assert_called_once_with(UPDATE_ATTRS)
 
 
-async def test_mug_update_multiple(ember_mug):
-    with patch.object(ember_mug, "_ensure_connection", AsyncMock()):
-        ember_mug.get_name = AsyncMock(return_value="name")
-        ember_mug.data.update_info = AsyncMock()
-        await ember_mug._update_multiple(("name",))
-        ember_mug.data.update_info.assert_called_once_with(name="name")
+async def test_mug_update_multiple(ember_mug: MockMug) -> None:
+    mock_get_name = AsyncMock(return_value="name")
+    mock_update_info = AsyncMock()
+
+    with patch.multiple(ember_mug, _ensure_connection=AsyncMock(), get_name=mock_get_name):
+        with patch.object(ember_mug.data, 'update_info', mock_update_info):
+            await ember_mug._update_multiple({"name"})
+            mock_get_name.assert_called_once()
+            mock_update_info.assert_called_once_with(name="name")
 
 
-async def test_mug_update_queued_attributes(ember_mug):
-    with patch.object(ember_mug, "_ensure_connection", AsyncMock()):
+async def test_mug_update_queued_attributes(ember_mug: MockMug) -> None:
+    mock_get_name = AsyncMock(return_value="name")
+    mock_update_info = AsyncMock()
+
+    with patch.multiple(ember_mug, _ensure_connection=AsyncMock(), get_name=mock_get_name):
         ember_mug._queued_updates = set()
         assert (await ember_mug.update_queued_attributes()) == []
-        ember_mug.get_name = AsyncMock(return_value="name")
-        ember_mug.data.update_info = AsyncMock()
-        ember_mug._queued_updates = {"name"}
-        await ember_mug.update_queued_attributes()
-        ember_mug.data.update_info.assert_called_once_with(name="name")
+        with patch.object(ember_mug.data, 'update_info', mock_update_info):
+            ember_mug._queued_updates = {"name"}
+            await ember_mug.update_queued_attributes()
+            mock_update_info.assert_called_once_with(name="name")
 
 
-def test_mug_notify_callback(ember_mug: EmberMug) -> None:
+def test_mug_notify_callback(ember_mug: MockMug) -> None:
     gatt_char = AsyncMock()
     ember_mug._notify_callback(gatt_char, bytearray(b"\x01"))
     ember_mug._notify_callback(gatt_char, bytearray(b"\x02"))